+++ ImageSlideshow/Classes/Core/FullScreenSlideshowViewController.swift
@@ -7,13 +7,14 @@
 
 import UIKit
 
+@objcMembers
 open class FullScreenSlideshowViewController: UIViewController {
 
     open var slideshow: ImageSlideshow = {
         let slideshow = ImageSlideshow()
         slideshow.zoomEnabled = true
         slideshow.contentScaleMode = UIViewContentMode.scaleAspectFit
-        slideshow.pageControlPosition = PageControlPosition.insideScrollView
+        slideshow.pageIndicatorPosition = PageIndicatorPosition(horizontal: .center, vertical: .bottom)
         // turns off the timer
         slideshow.slideshowInterval = 0
         slideshow.autoresizingMask = [UIViewAutoresizing.flexibleWidth, UIViewAutoresizing.flexibleHeight]
@@ -23,6 +24,9 @@ open class FullScreenSlideshowViewController: UIViewController {
 
     /// Close button 
     open var closeButton = UIButton()
+
+    /// Close button frame
+    open var closeButtonFrame: CGRect?
 
     /// Closure called on page selection
     open var pageSelected: ((_ page: Int) -> Void)?
@@ -49,21 +53,16 @@ open class FullScreenSlideshowViewController: UIViewController {
         super.viewDidLoad()
 
         view.backgroundColor = backgroundColor
-
-        // slideshow view configuration
-        slideshow.frame = view.frame
         slideshow.backgroundColor = backgroundColor
 
         if let inputs = inputs {
             slideshow.setImageInputs(inputs)
         }
 
-        slideshow.frame = view.frame
         view.addSubview(slideshow)
 
         // close button configuration
-        closeButton.frame = CGRect(x: 10, y: 20, width: 40, height: 40)
-        closeButton.setImage(UIImage(named: "Frameworks/ImageSlideshow.framework/ImageSlideshow.bundle/ic_cross_white@2x"), for: UIControlState())
+        closeButton.setImage(UIImage(named: "ic_cross_white", in: Bundle(for: type(of: self)), compatibleWith: nil), for: UIControlState())
         closeButton.addTarget(self, action: #selector(FullScreenSlideshowViewController.close), for: UIControlEvents.touchUpInside)
         view.addSubview(closeButton)
     }
@@ -81,11 +80,28 @@ open class FullScreenSlideshowViewController: UIViewController {
         }
     }
 
+    override open func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        
+        slideshow.slideshowItems.forEach { $0.cancelPendingLoad() }
+    }
+
     open override func viewDidLayoutSubviews() {
+        if !isBeingDismissed {
+            let safeAreaInsets: UIEdgeInsets
+            if #available(iOS 11.0, *) {
+                safeAreaInsets = view.safeAreaInsets
+            } else {
+                safeAreaInsets = UIEdgeInsets.zero
+            }
+            
+            closeButton.frame = closeButtonFrame ?? CGRect(x: max(10, safeAreaInsets.left), y: max(10, safeAreaInsets.top), width: 40, height: 40)
+        }
+
         slideshow.frame = view.frame
     }
 
-    func close() {
+    @objc func close() {
         // if pageSelected closure set, send call it with current page
         if let pageSelected = pageSelected {
             pageSelected(slideshow.currentPage)