+++ ImageSlideshow/Classes/Core/ImageSlideshow.swift
@@ -19,17 +19,6 @@ public enum PageControlPosition {
     case insideScrollView
     case underScrollView
     case custom(padding: CGFloat)
-
-    var bottomPadding: CGFloat {
-        switch self {
-        case .hidden, .insideScrollView:
-            return 0.0
-        case .underScrollView:
-            return 30.0
-        case .custom(let padding):
-            return padding
-        }
-    }
 }
 
 /// Used to represent image preload strategy
@@ -42,35 +31,67 @@ public enum ImagePreload {
 }
 
 /// Main view containing the Image Slideshow
+@objcMembers
 open class ImageSlideshow: UIView {
 
     /// Scroll View to wrap the slideshow
     open let scrollView = UIScrollView()
 
     /// Page Control shown in the slideshow
-    open let pageControl = UIPageControl()
+    @available(*, deprecated, message: "Use pageIndicator.view instead")
+    open var pageControl: UIPageControl {
+        if let pageIndicator = pageIndicator as? UIPageControl {
+            return pageIndicator
+        }
+        fatalError("pageIndicator is not an instance of UIPageControl")
+    }
 
     /// Activity indicator shown when loading image
     open var activityIndicator: ActivityIndicatorFactory? {
         didSet {
-            self.reloadScrollView()
+            reloadScrollView()
+        }
+    }
+
+    open var pageIndicator: PageIndicatorView? = UIPageControl() {
+        didSet {
+            oldValue?.view.removeFromSuperview()
+            if let pageIndicator = pageIndicator {
+                addSubview(pageIndicator.view)
+            }
+            setNeedsLayout()
+        }
+    }
+
+    open var pageIndicatorPosition: PageIndicatorPosition = PageIndicatorPosition() {
+        didSet {
+            setNeedsLayout()
         }
     }
 
     // MARK: - State properties
 
     /// Page control position
+    @available(*, deprecated, message: "Use pageIndicatorPosition instead")
     open var pageControlPosition = PageControlPosition.insideScrollView {
         didSet {
-            setNeedsLayout()
+            pageIndicator = UIPageControl()
+            switch pageControlPosition {
+            case .hidden:
+                pageIndicator = nil
+            case .insideScrollView:
+                pageIndicatorPosition = PageIndicatorPosition(vertical: .bottom)
+            case .underScrollView:
+                pageIndicatorPosition = PageIndicatorPosition(vertical: .under)
+            case .custom(let padding):
+                pageIndicatorPosition = PageIndicatorPosition(vertical: .customUnder(padding: padding-30))
+            }
         }
     }
 
     /// Current page
     open fileprivate(set) var currentPage: Int = 0 {
         didSet {
-            pageControl.currentPage = currentPage
-
             if oldValue != currentPage {
                 currentPageChanged?(currentPage)
             }
@@ -112,8 +133,8 @@ open class ImageSlideshow: UIView {
     /// Enables/disables infinite scrolling between images
     open var circular = true {
         didSet {
-            if self.images.count > 0 {
-                self.setImageInputs(self.images)
+            if images.count > 0 {
+                setImageInputs(images)
             }
         }
     }
@@ -121,22 +142,29 @@ open class ImageSlideshow: UIView {
     /// Enables/disables user interactions
     open var draggingEnabled = true {
         didSet {
-            self.scrollView.isUserInteractionEnabled = draggingEnabled
+            scrollView.isUserInteractionEnabled = draggingEnabled
         }
     }
 
     /// Enables/disables zoom
     open var zoomEnabled = false {
         didSet {
-            self.reloadScrollView()
+            reloadScrollView()
+        }
+    }
+    
+    /// Maximum zoom scale
+    open var maximumScale: CGFloat = 2.0 {
+        didSet {
+            reloadScrollView()
         }
     }
 
     /// Image change interval, zero stops the auto-scrolling
     open var slideshowInterval = 0.0 {
         didSet {
-            self.slideshowTimer?.invalidate()
-            self.slideshowTimer = nil
+            slideshowTimer?.invalidate()
+            slideshowTimer = nil
             setTimerIfNeeded()
         }
     }
@@ -161,6 +189,10 @@ open class ImageSlideshow: UIView {
 
     /// Transitioning delegate to manage the transition to full screen controller
     open fileprivate(set) var slideshowTransitioningDelegate: ZoomAnimatedTransitioningDelegate?
+    
+    var primaryVisiblePage: Int {
+        return scrollView.frame.size.width > 0 ? Int(scrollView.contentOffset.x + scrollView.frame.size.width / 2) / Int(scrollView.frame.size.width) : 0
+    }
 
     // MARK: - Life cycle
 
@@ -189,11 +221,19 @@ open class ImageSlideshow: UIView {
         scrollView.bounces = true
         scrollView.showsHorizontalScrollIndicator = false
         scrollView.showsVerticalScrollIndicator = false
-        scrollView.autoresizingMask = self.autoresizingMask
+        scrollView.autoresizingMask = autoresizingMask
+        if #available(iOS 11.0, *) {
+            scrollView.contentInsetAdjustmentBehavior = .never
+        }
         addSubview(scrollView)
 
-        addSubview(pageControl)
-        pageControl.addTarget(self, action: #selector(pageControlValueChanged), for: .valueChanged)
+        if let pageIndicator = pageIndicator {
+            addSubview(pageIndicator.view)
+        }
+        
+        if let pageIndicator = pageIndicator as? UIControl {
+            pageIndicator.addTarget(self, action: #selector(pageControlValueChanged), for: .valueChanged)
+        }
 
         setTimerIfNeeded()
         layoutScrollView()
@@ -201,7 +241,7 @@ open class ImageSlideshow: UIView {
 
     open override func removeFromSuperview() {
         super.removeFromSuperview()
-        self.pauseTimer()
+        pauseTimer()
     }
 
     open override func layoutSubviews() {
@@ -215,44 +255,49 @@ open class ImageSlideshow: UIView {
     }
 
     open func layoutPageControl() {
-        if case .hidden = self.pageControlPosition {
-            pageControl.isHidden = true
-        } else {
-            pageControl.isHidden = self.images.count < 2
-        }
+        if let pageIndicatorView = pageIndicator?.view {
+            pageIndicatorView.isHidden = images.count < 2
 
-        pageControl.frame = CGRect(x: 0, y: 0, width: frame.size.width, height: 10)
-        pageControl.center = CGPoint(x: frame.size.width / 2, y: frame.size.height - 12.0)
+            var edgeInsets: UIEdgeInsets = UIEdgeInsets.zero
+            if #available(iOS 11.0, *) {
+                edgeInsets = safeAreaInsets
+            }
+
+            pageIndicatorView.sizeToFit()
+            pageIndicatorView.frame = pageIndicatorPosition.indicatorFrame(for: frame, indicatorSize: pageIndicatorView.frame.size, edgeInsets: edgeInsets)
+        }
     }
 
     /// updates frame of the scroll view and its inner items
     func layoutScrollView() {
-        let scrollViewBottomPadding: CGFloat = pageControlPosition.bottomPadding
+        let pageIndicatorViewSize = pageIndicator?.view.frame.size
+        let scrollViewBottomPadding = pageIndicatorViewSize.flatMap { pageIndicatorPosition.underPadding(for: $0) } ?? 0
+
         scrollView.frame = CGRect(x: 0, y: 0, width: frame.size.width, height: frame.size.height - scrollViewBottomPadding)
         scrollView.contentSize = CGSize(width: scrollView.frame.size.width * CGFloat(scrollViewImages.count), height: scrollView.frame.size.height)
 
-        for (index, view) in self.slideshowItems.enumerated() {
+        for (index, view) in slideshowItems.enumerated() {
             if !view.zoomInInitially {
                 view.zoomOut()
             }
             view.frame = CGRect(x: scrollView.frame.size.width * CGFloat(index), y: 0, width: scrollView.frame.size.width, height: scrollView.frame.size.height)
         }
 
-        setCurrentPage(currentPage, animated: false)
+        setScrollViewPage(scrollViewPage, animated: false)
     }
 
     /// reloads scroll view with latest slideshow items
     func reloadScrollView() {
         // remove previous slideshow items
-        for view in self.slideshowItems {
+        for view in slideshowItems {
             view.removeFromSuperview()
         }
-        self.slideshowItems = []
+        slideshowItems = []
 
         var i = 0
         for image in scrollViewImages {
-            let item = ImageSlideshowItem(image: image, zoomEnabled: self.zoomEnabled, activityIndicator: self.activityIndicator?.create(), fallbackImage: fallbackImage, fallbackScaleMode: self.fallbackScaleMode)
-            item.imageView.contentMode = self.contentScaleMode
+            let item = ImageSlideshowItem(image: image, zoomEnabled: zoomEnabled, activityIndicator: activityIndicator?.create(), maximumScale: maximumScale, fallbackImage: fallbackImage, fallbackScaleMode: self.fallbackScaleMode)
+            item.imageView.contentMode = contentScaleMode
             slideshowItems.append(item)
             scrollView.addSubview(item)
             i += 1
@@ -273,7 +318,7 @@ open class ImageSlideshow: UIView {
 
         for i in 0..<totalCount {
             let item = slideshowItems[i]
-            switch self.preload {
+            switch preload {
             case .all:
                 item.loadImage()
             case .fixed(let offset):
@@ -294,8 +339,8 @@ open class ImageSlideshow: UIView {
      - parameter inputs: Array of InputSource instances.
      */
     open func setImageInputs(_ inputs: [InputSource]) {
-        self.images = inputs
-        self.pageControl.numberOfPages = inputs.count
+        images = inputs
+        pageIndicator?.numberOfPages = inputs.count
 
         // in circular mode we add dummy first and last image to enable smooth scrolling
         if circular && images.count > 1 {
@@ -309,9 +354,9 @@ open class ImageSlideshow: UIView {
                 scImages.append(first)
             }
 
-            self.scrollViewImages = scImages
+            scrollViewImages = scImages
         } else {
-            self.scrollViewImages = images
+            scrollViewImages = images
         }
 
         reloadScrollView()
@@ -333,7 +378,7 @@ open class ImageSlideshow: UIView {
             pageOffset += 1
         }
 
-        self.setScrollViewPage(pageOffset, animated: animated)
+        setScrollViewPage(pageOffset, animated: animated)
     }
 
     /**
@@ -343,8 +388,8 @@ open class ImageSlideshow: UIView {
      */
     open func setScrollViewPage(_ newScrollViewPage: Int, animated: Bool) {
         if scrollViewPage < scrollViewImages.count {
-            self.scrollView.scrollRectToVisible(CGRect(x: scrollView.frame.size.width * CGFloat(newScrollViewPage), y: 0, width: scrollView.frame.size.width, height: scrollView.frame.size.height), animated: animated)
-            self.setCurrentPageForScrollViewPage(newScrollViewPage)
+            scrollView.scrollRectToVisible(CGRect(x: scrollView.frame.size.width * CGFloat(newScrollViewPage), y: 0, width: scrollView.frame.size.width, height: scrollView.frame.size.height), animated: animated)
+            setCurrentPageForScrollViewPage(newScrollViewPage)
         }
     }
 
@@ -354,15 +399,15 @@ open class ImageSlideshow: UIView {
         }
     }
 
-    func slideshowTick(_ timer: Timer) {
-        let page = Int(scrollView.contentOffset.x / scrollView.frame.size.width)
+    @objc func slideshowTick(_ timer: Timer) {
+        let page = scrollView.frame.size.width > 0 ? Int(scrollView.contentOffset.x / scrollView.frame.size.width) : 0
         var nextPage = page + 1
 
         if !circular && page == scrollViewImages.count - 1 {
             nextPage = 0
         }
 
-        self.setScrollViewPage(nextPage, animated: true)
+        setScrollViewPage(nextPage, animated: true)
     }
 
     fileprivate func setCurrentPageForScrollViewPage(_ page: Int) {
@@ -377,19 +422,22 @@ open class ImageSlideshow: UIView {
             loadImages(for: page)
         }
         scrollViewPage = page
-
+        currentPage = currentPageForScrollViewPage(page)
+    }
+    
+    fileprivate func currentPageForScrollViewPage(_ page: Int) -> Int {
         if circular {
             if page == 0 {
                 // first page contains the last image
-                currentPage = Int(images.count) - 1
+                return Int(images.count) - 1
             } else if page == scrollViewImages.count - 1 {
                 // last page contains the first image
-                currentPage = 0
+                return 0
             } else {
-                currentPage = page - 1
+                return page - 1
             }
         } else {
-            currentPage = page
+            return page
         }
     }
 
@@ -406,12 +454,12 @@ open class ImageSlideshow: UIView {
 
     @available(*, deprecated, message: "use pauseTimer instead")
     open func pauseTimerIfNeeded() {
-        self.pauseTimer()
+        pauseTimer()
     }
 
     @available(*, deprecated, message: "use unpauseTimer instead")
     open func unpauseTimerIfNeeded() {
-        self.unpauseTimer()
+        unpauseTimer()
     }
 
     /**
@@ -422,12 +470,12 @@ open class ImageSlideshow: UIView {
     @discardableResult
     open func presentFullScreenController(from controller: UIViewController) -> FullScreenSlideshowViewController {
         let fullscreen = FullScreenSlideshowViewController()
-        fullscreen.pageSelected = {(page: Int) in
-            self.setCurrentPage(page, animated: false)
+        fullscreen.pageSelected = {[weak self] (page: Int) in
+            self?.setCurrentPage(page, animated: false)
         }
 
-        fullscreen.initialPage = self.currentPage
-        fullscreen.inputs = self.images
+        fullscreen.initialPage = currentPage
+        fullscreen.inputs = images
         slideshowTransitioningDelegate = ZoomAnimatedTransitioningDelegate(slideshowView: self, slideshowController: fullscreen)
         fullscreen.transitioningDelegate = slideshowTransitioningDelegate
         controller.present(fullscreen, animated: true, completion: nil)
@@ -436,7 +484,9 @@ open class ImageSlideshow: UIView {
     }
 
     @objc private func pageControlValueChanged() {
-        self.setCurrentPage(pageControl.currentPage, animated: true)
+        if let currentPage = pageIndicator?.page {
+            setCurrentPage(currentPage, animated: true)
+        }
     }
 }
 
@@ -453,8 +503,7 @@ extension ImageSlideshow: UIScrollViewDelegate {
     }
 
     open func scrollViewDidEndDecelerating(_ scrollView: UIScrollView) {
-        let page = Int(scrollView.contentOffset.x) / Int(scrollView.frame.size.width)
-        setCurrentPageForScrollViewPage(page)
+        setCurrentPageForScrollViewPage(primaryVisiblePage)
         didEndDecelerating?()
     }
 
@@ -468,5 +517,7 @@ extension ImageSlideshow: UIScrollViewDelegate {
                 scrollView.contentOffset = CGPoint(x: scrollView.contentOffset.x + regularContentOffset, y: 0)
             }
         }
+
+        pageIndicator?.page = currentPageForScrollViewPage(primaryVisiblePage)
     }
 }